+++ spec/integration/resource/api500/synergy/switch/create_spec.rb
@@ -1,4 +1,4 @@
-# (C) Copyright 2016 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2017 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # You may not use this file except in compliance with the License.
@@ -11,8 +11,8 @@
 
 require 'spec_helper'
 
-klass = OneviewSDK::API300::C7000::LogicalDownlink
+klass = OneviewSDK::API500::Synergy::Switch
 RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
-  let(:current_client) { $client_300 }
-  include_examples 'LogicalDownlinkCreateExample', 'integration api300 context', true
+  let(:current_client) { $client_500_synergy }
+  include_examples 'SwitchCreateExample', 'integration api500 context'
 end