+++ spec/integration/shared_examples/sas_logical_interconnect/create.rb
@@ -1,4 +1,4 @@
-# (C) Copyright 2016 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2017 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # You may not use this file except in compliance with the License.
@@ -9,10 +9,13 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-require 'spec_helper'
+RSpec.shared_examples 'SASLogicalInterconnectCreateExample' do |context_name|
+  include_context context_name
 
-klass = OneviewSDK::API300::C7000::LogicalDownlink
-RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
-  let(:current_client) { $client_300 }
-  include_examples 'LogicalDownlinkCreateExample', 'integration api300 context', true
+  describe '#create' do
+    it 'self raises MethodUnavailable' do
+      item = described_class.new(current_client)
+      expect { item.create }.to raise_error(/The method #create is unavailable for this resource/)
+    end
+  end
 end