+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased Changes
 
-## Suggested release: v4.3.0
+## Suggested release: v5.0.0
 
 #### Features supported
 This release adds support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
@@ -8,11 +8,17 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - FC Network
    - FCoE Network
    - Interconnect
+   - Logical Downlink
    - Logical Interconnect Group
    - Logical Switch Group
    - Network Set
+   - SAS Logical Interconnect
    - SAS Logical Interconnect Group
    - Scopes
+   - Switches
+
+#### Breaking changes
+1. In the Switch resource, the `statistics` method was fixed, a non-existent endpoint was removed `/rest/switches/statistics/{portName}/subport/{subportNumber}`. Now the method only get statistics of a switch or a specific port.
 
 # v4.2.0